+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/features/storage/serialization/logs/mapping/collection/LogCollectionConverter.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.features.storage.serialization.logs.mapping;
+package co.elastic.apm.android.sdk.internal.features.storage.serialization.logs.mapping.collection;
 
 import java.util.ArrayList;
 import java.util.HashMap;
